Moved sub-classes of "FileDialog", "OpenFileDialog" and "SaveFileDialog" classes to their own files
