+++ src/System.Windows.Forms/src/System/Windows/Forms/SaveFileDialog.cs
@@ -6,7 +6,6 @@
 
 using System.ComponentModel;
 using System.IO;
-using System.Runtime.InteropServices;
 using static Interop;
 using static Interop.Shell32;
 
@@ -21,7 +20,7 @@ namespace System.Windows.Forms
     Designer("System.Windows.Forms.Design.SaveFileDialogDesigner, " + AssemblyRef.SystemDesign)]
     [SRDescription(nameof(SR.DescriptionSaveFileDialog))
     ]
-    public sealed class SaveFileDialog : FileDialog
+    public sealed partial class SaveFileDialog : FileDialog
     {
         /// <summary>
         ///  Gets or sets a value indicating whether the dialog box prompts the user for
@@ -153,20 +152,5 @@ namespace System.Windows.Forms
         }
 
         private protected override IFileDialog CreateVistaDialog() => new NativeFileSaveDialog();
-
-        [ComImport]
-        [Guid("84bccd23-5fde-4cdb-aea4-af64b83d78ab")]
-        [CoClass(typeof(FileSaveDialogRCW))]
-        private interface NativeFileSaveDialog : IFileSaveDialog
-        {
-        }
-
-        [ComImport]
-        [ClassInterface(ClassInterfaceType.None)]
-        [TypeLibType(TypeLibTypeFlags.FCanCreate)]
-        [Guid("C0B4E2F3-BA21-4773-8DBA-335EC946EB8B")]
-        private class FileSaveDialogRCW
-        {
-        }
     }
 }
